// src/app/interfaces/common/theme-category.interface.ts
export interface ThemeCategory {
  _id?: string;
  readOnly?: boolean;
  name?: string;
  slug?: string;
  image?: string;
  description?: string;
  serial?: number;
  createdAt?: Date;
  updatedAt?: Date;
  status?: string;
  select?: boolean;
  featureStatus?: boolean;
  featureImage?: string;

}